Restructure components and screens directory
