+++ components/Buttons/ChangeBtns/GenderBtn/GenderPresenter.js
@@ -1,15 +1,15 @@
 import React, { memo } from 'react';
-import ToggleButton from 'react-native-toggle-element';
 import { Image } from 'react-native';
+import ToggleButton from 'react-native-toggle-element';
 
-const TwoPeopleBtn = memo((props) => (
+const GenderBtn = memo((props) => (
   <ToggleButton
     onPress={props.onPress}
     value={props.value}
     onToggle={props.onToggle}
     thumbActiveComponent={
       <Image
-        source={require('../../../assets/images/two_people.png')}
+        source={require('../../../../assets/images/man.png')}
         style={{
           width: '100%',
           height: '100%',
@@ -19,7 +19,7 @@ const TwoPeopleBtn = memo((props) => (
     }
     thumbInActiveComponent={
       <Image
-        source={require('../../../assets/images/one_person.png')}
+        source={require('../../../../assets/images/woman.png')}
         style={{
           width: '100%',
           height: '100%',
@@ -28,8 +28,8 @@ const TwoPeopleBtn = memo((props) => (
       />
     }
     thumbButton={{
-      width: 100,
-      height: 100,
+      width: 50,
+      height: 50,
       radius: 0,
     }}
     trackBar={{
@@ -37,13 +37,13 @@ const TwoPeopleBtn = memo((props) => (
       inActiveBackgroundColor: 'transparent',
       borderActiveColor: 'white',
       borderInActiveColor: 'white',
-      width: 100,
-      height: 100,
+      width: 50,
+      height: 50,
     }}
     trackBarStyle={{
       opacity: 0.9,
     }}
-  ></ToggleButton>
+  />
 ));
 
-export default TwoPeopleBtn;
+export default GenderBtn;
