+++ components/Buttons/MainScreenBtns/GetPhotoBtn/GetPhotoContainer.js
