+++ components/Buttons/MainScreenBtns/GetPhotoBtn/GetPhotoPresenter.js
