+++ components/Buttons/MainScreenBtns/NoticeBtns/NoticeCancelBtn/NoticeCancelPresenter.js
