+++ components/Buttons/MainScreenBtns/NoticeBtns/NoticeNeverBtn/NoticeNeverPresenter.js
