+++ components/Buttons/MainScreenBtns/SwitchCameraBtn/SwitchCameraPresenter.js
