+++ components/Buttons/MainScreenBtns/TakePhotoBtn/TakePhotoContainer.js
