+++ components/Buttons/MainScreenBtns/TakePhotoBtn/TakePhotoPresenter.js
