+++ components/ProgressBar/progressBarMain.js
